split Object data extraction from view
